LayoutTests/media/media-source/media-source-real-webm-no-default-duration-expected.txt


EVENT(sourceopen)
EVENT(updateend)
EVENT(updateend)
EXPECTED (sourceBuffer.buffered.length == '1') OK
EXPECTED (sourceBuffer.buffered.start(0) == '0') OK
EXPECTED (sourceBuffer.buffered.end(0) == '4') OK
END OF TEST
